Gearing towards a text-based GUI for passwords.
This commit contributes a bunch of new ideas, some of which were months old and
not committed. In particular, there is now a method to create a system of dicts
which are traversible by POSIX-style paths, a couple of methods to detect the
local console width and format the output of a directory listing appropriately,
and a file for creating a generic sh-style read-eval-print loop with high-level
commands. (Although it at this stage lacks variables of any sort and any sort
of expression chaining.)